refactor: ester egg page 생성
related to: #160
