+++ src/Onboarding/index.tsx
@@ -1,12 +1,12 @@
 import { css } from '@emotion/react';
 
+import Hyeonjin from '../esterEgg/components/Hyeonjin';
 import Frame1 from './components/frame/Frame1';
 import Frame2 from './components/frame/Frame2';
 import Frame3 from './components/frame/Frame3';
 import Frame4 from './components/frame/Frame4';
 import Frame5 from './components/frame/Frame5';
 import Frame6 from './components/frame/Frame6';
-import Hyeonjin from './components/frame/Hyeonjin';
 
 const Onboarding = () => {
   return (
